requirements.txt
Flask==2.3.0
Flask-JWT-Extended==4.4.4
Flask_SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
openai==1.52.0
python-dotenv==1.0.1
requests==2.31.0
werkzeug==2.3.0
pymysql==1.0.2
cos-python-sdk-v5==1.9.32
flask_cors==3.0.10
gunicorn==23.0.0
aliyun_python_sdk_core==2.13.36
aliyun_python_sdk_ecs==4.24.62
flask_migrate==4.0.7
flask-admin==1.6.1
flask-login==0.6.3
flask-wtf==1.2.2